+++ click_extra/parameters.py
@@ -1660,13 +1660,14 @@ class ShowParamsOption(ExtraOption, ParamStructure):
     def __init__(
         self,
         param_decls: Sequence[str] | None = None,
-        is_flag=True,
-        expose_value=False,
-        is_eager=True,
-        help=_(
+        *,
+        is_flag: bool = True,
+        expose_value: bool = False,
+        is_eager: bool = True,
+        help: str = _(
             "Show all CLI parameters, their provenance, defaults and value, then exit.",
         ),
-        **kwargs,
+        **kwargs: Any,
     ) -> None:
         if not param_decls:
             param_decls = ("--params",)